Add domain, persistence and services locations.
