+++ MonitoringService/Domain/Models/SpecDetails.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 
-namespace MonitoringService
+namespace MonitoringService.Domain.Models
 {
     public class SpecDetails
     {
