+++ MonitoringService/Persistence/ApplicationContext.cs
@@ -4,8 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
+using MonitoringService.Domain.Models;
 
-namespace MonitoringService 
+namespace MonitoringService.Persistence
 {
     public class ApplicationContext : DbContext
     {
